Added user authentication and registration, this closes #1
Added routes:
- POST - auth/login
- POST - auth/register
- DEL - auth/logout

The login route returns the user data and a cookie with a signed jwt
token. Both the jwt token and the cookie have a expiration time of 1
hour. The cookie is then parsed in the server.js of sapper with the
cookie-parser module and set the user as authenticated in the sapper
session. At logout the api returns a expired cookie by setting the
expiration date of -1, this leads to the deletion of cookie at the
client side.